+++ py_gql/sdl/__init__.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+"""
+Utilities to work with the GraphQL schema definition language (SDL).
+"""
+
+# flake8: noqa
+
+from .schema_directives import SchemaDirective, apply_schema_directives
+from .schema_from_ast import build_schema, extend_schema
+
+__all__ = (
+    "build_schema",
+    "extend_schema",
+    "SchemaDirective",
+    "apply_schema_directives",
+)